Add files via upload
Bash script to batch extract page titles from TIFF headers with their filenames and put them as columns in a spreadsheet.
Requirements: ExifTool; connexion to network storage, e.g. NetApp
Once the scanning of the directory and reading of the image files have been completed, the CSV will be complete with page filenames extracted from the source path of the image file and page titles extracted from the TIFF header. This CSV may be imported into a Google Sheet and/or submitted to the Samvera team.